proto: change Flight.proto::doget to accept stream
For many cases passing a ticket into doget method is OK, but if the schema is large enough,
the ticket becomes heavy. It's not a big problem itself though. It does become a problem,
when used gRPC connection is overloaded. The default mechanism to handle that is to put waiting
requests into a queue and take them one-by-one once the connection is ready. If each ticket
is heavy, and there are many of such incoming requests, then the queue starts utilizing noticeable
amount of memory.

The patch allows to customize the discribed behavior. Instead of consuming a potentially heavy ticket,
doget accepts a stream over the ticket. When needed the stream can be customized such way
that life time of its element is controlled outside of the stream (e. g. it can be pinned to
the parent future to be expired when the future drops). Moreover, the stream itself may become lightweight
because it isn't required to consume the ticket anymore, instead it may contain a weak pointer,
receiver, or use other technique of promising.